pyproject.toml
[project]
name = "dj-mypypi2"
description = "Basic PyPI server for Django projects"
authors = [
    { name = "Karel Antonio Verdecia Ortiz", email = "kverdecia@gmail.com" }
]
maintainers = [
    { name = "Karel Antonio Verdecia Ortiz", email = "kverdecia@gmail.com" }
]
license = "MIT"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "cuid>=0.4",
    "django>=5.2.3",
    "factory-boy>=3.3.3",
    "python-dotenv>=1.1.0",
    "requests>=2.32.4",
    "rich>=14.0.0",
]
dynamic = ["version"]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Framework :: Django :: 5.0",
    "Intended Audience :: Developers",
    "Natural Language :: English",
    "Programming Language :: Python :: 3.12",
]

[project.urls]
Homepage = "https://github.com/kverdecia/dj-mypypi2"
Repository = "https://github.com/kverdecia/dj-mypypi2"
Issues = "https://github.com/kverdecia/dj-mypypi2/issuess"

[dependency-groups]
dev = [
    "codecov>=2.0.0",
    "coverage==4.4.1",
    "dj-inmemorystorage>=2.1.0",
    "django-stubs>=5.2.1",
    "factory-boy>=3.3.3",
    "flake8>=2.1.0",
    "invoke>=2.2.0",
    "mock>=1.0.1",
    "mypy>=1.16.1",
    "tox>=1.7.0",
    "twine>=6.1.0",
    "wheel>=0.45.1",
    "zest-releaser>=9.6.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/djmypypi2"]

[tool.hatch.version]
path = "./src/djmypypi2/__init__.py"
